Move handlers to own class
